Delete BookHelper and move methods to BookRepository, BookService
