+++ InteractionController.cs
@@ -70,6 +70,12 @@ public class InteractionController(IBookService bookService)
             .ForEach(Console.WriteLine);
     }
 
+    // from BookHelper
+    static void PrintUpdatedBookProperties(string updatedBook)
+    {
+        Console.WriteLine("\nUpdated. New properties are as follows: ");
+        Console.WriteLine(ConvertLineToPropertiesList(updatedBook));
+    }
     string ConvertLineToReadableInfo(string? author, List<Book> books)
     {
         return $"Author: {author}\n" + string.Join("\n", books.Select(ConvertLineToReadableInfo));
@@ -79,5 +85,59 @@ public class InteractionController(IBookService bookService)
     {
         return $"{index + 1}. Title: {book.Title}. Author: {book.Author}. Year published: {book.PublishYear}";
     }
-    
+
+    // from BookHelper
+    public static List<Book> GetUsersListOfBooks()
+    {
+        bool run = true;
+        List<Book> books = new List<Book>();
+
+        do
+        {
+            books.Add(GetUserInputAsBook());
+            Console.WriteLine("\nDo you wish to add another book? Y/N");
+            string userResponse = Console.ReadLine();
+            if (userResponse == "y")
+            {
+                continue;
+            }
+            else if (userResponse == "n")
+            {
+                break;
+            }
+            else
+            {
+                Console.WriteLine("Invalid response, please try again.");
+            };
+
+        } while (run);
+
+        return books;
+    }
+
+    // from BookHelper
+    static Book GetUserInputAsBook()
+    {
+        // dummy pid
+        int pid = 0;
+
+        Console.WriteLine("\nPlease enter the book title: ");
+        string title = Console.ReadLine();
+
+        Console.WriteLine("\nPlease enter the author name: ");
+        string author = Console.ReadLine();
+
+        Console.WriteLine("\nPlease enter the publish date: ");
+        int publishDate = int.Parse(Console.ReadLine());
+
+        Book book = new Book();
+        book.Id = pid;
+        book.Title = title;
+        book.Author = author;
+        book.PublishYear = publishDate;
+
+        return book;
+    }
+}
+
 }
